requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
attrs==25.4.0
beautifulsoup4==4.12.3
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.1
discord.py==2.6.4
distlib==0.3.9
filelock==3.17.0
frozenlist==1.8.0
h11==0.16.0
idna==3.10
multidict==6.7.0
numpy==2.3.3
outcome==1.3.0.post0
pandas==2.3.3
platformdirs==4.3.6
propcache==0.4.1
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.3
selenium==4.36.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.31.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.29.1
websocket-client==1.9.0
wsproto==1.2.0
xlrd==2.0.2
yarl==1.22.0
